删除属性组时同时删除描述 删除属性组时检查是否有属性在使用 删除属性时同时删除描述

+++ beike/Admin/Repositories/AttributeGroupRepo.php
@@ -69,6 +69,10 @@ class AttributeGroupRepo
     public static function delete($id)
     {
         $group = AttributeGroup::query()->findOrFail($id);
+        if ($group->attributes->count()) {
+            throw New \Exception(trans('admin/attribute_groups.error_cannot_delete_attribute_used', ['attributes' => implode(', ', $group->attributes->pluck('id')->toArray())]));
+        }
+        $group->descriptions()->delete();
         $group->delete();
     }
 }
